Move ZvmNode from common to configparser
