+++ test/unit/test_containerquery.py
@@ -11,8 +11,12 @@ from swift.container.server import ContainerController
 from test.unit import FakeLogger, create_tar, trim
 from test.unit.test_proxyquery import ZEROVM_DEFAULT_MOCK
 from zerocloud import objectquery
-from zerocloud.common import ZvmNode, parse_location, ACCESS_READABLE, \
-    ACCESS_WRITABLE, NodeEncoder, json
+from zerocloud.common import parse_location
+from zerocloud.common import ACCESS_READABLE
+from zerocloud.common import ACCESS_WRITABLE
+from zerocloud.common import json
+from zerocloud.configparser import NodeEncoder
+from zerocloud.configparser import ZvmNode
 from zerocloud.thread_pool import Zuid
 from eventlet.wsgi import Input
